requirements.txt
# Telegram Bot
aiogram==3.13.1

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Config & Validation
pydantic>=2.4.1,<2.10
pydantic-settings>=2.6.0,<3.0

# OpenAI
openai>=1.50.0

# Document Parsing
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# HTTP & Utils
aiohttp==3.9.1
python-dotenv==1.0.0

# PDF (future)
pypdf==3.17.4
pdfplumber==0.10.3

# Testing
pytest>=7.0
pytest-asyncio>=0.21
